+++ core/types/cover_art.go
@@ -2,5 +2,5 @@ package types
 
 type CoverArt struct {
 	Id   string `xml:"id,attr" json:"id"`
-	Size int64  `xml:"size,attr" json:"size"`
+	Size int    `xml:"size,attr" json:"size"`
 }
